moved images and fonts
created assets folder and moved binary files there
